api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
boto3==1.29.6
aioboto3==12.1.0
structlog==23.2.0
python-json-logger==2.0.7
